refactor: 패키지 구조 변경
BallCounter, StrikeCounter 객체가 사라짐에 따라 패키지 구조를 변경하였습니다
