+++ src/main/java/baseball/domain/SingleNumber.java
@@ -1,4 +1,4 @@
-package baseball.domain.number;
+package baseball.domain;
 
 import java.util.List;
 import java.util.Objects;
