+++ src/test/java/baseball/domain/count/CountResultTest.java
@@ -1,5 +1,6 @@
 package baseball.domain.count;
 
+import baseball.domain.CountResult;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
